+++ examples/send_e2e_text.rs
@@ -3,7 +3,7 @@ use std::process;
 use docopt::Docopt;
 use threema_gateway::{ApiBuilder, RecipientKey};
 
-const USAGE: &'static str = "
+const USAGE: &str = "
 Usage: send_e2e_text [options] <from> <to> <secret> <private-key> <text>...
 
 Options:
@@ -36,7 +36,7 @@ fn main() {
     });
 
     // Encrypt and send
-    let recipient_key = RecipientKey::from_str(&public_key).unwrap_or_else(|e| {
+    let recipient_key: RecipientKey = public_key.parse().unwrap_or_else(|e| {
         println!("{}", e);
         process::exit(1);
     });