+++ src/lib.rs
@@ -40,16 +40,16 @@
 //!
 //! // Create E2eApi instance
 //! let api = ApiBuilder::new(from, secret)
-//!                      .with_private_key_str(private_key)
-//!                      .and_then(|builder| builder.into_e2e())
-//!                      .unwrap();
+//!     .with_private_key_str(private_key)
+//!     .and_then(|builder| builder.into_e2e())
+//!     .unwrap();
 //!
 //! // Fetch public key
 //! // Note: In a real application, you should cache the public key
 //! let public_key = api.lookup_pubkey(to).unwrap();
 //!
 //! // Encrypt
-//! let recipient_key = RecipientKey::from_str(&public_key).unwrap();
+//! let recipient_key: RecipientKey = public_key.parse().unwrap();
 //! let encrypted = api.encrypt_text_msg(text, &recipient_key);
 //!
 //! // Send
@@ -82,7 +82,7 @@ pub use crate::crypto::{EncryptedMessage, RecipientKey};
 pub use crate::lookup::{Capabilities, LookupCriterion};
 pub use crate::types::{BlobId, MessageType};
 
-const MSGAPI_URL: &'static str = "https://msgapi.threema.ch";
+const MSGAPI_URL: &str = "https://msgapi.threema.ch";
 
 #[cfg(test)]
 mod tests {